Collapse constellation/script/layout creation abstraction into script crate.
